requirements.txt
streamlit
groq
careerjet-api
python-dotenv
requests
